Move tests to different namespace
