+++ composer.json
@@ -31,7 +31,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-          "Zamzar\\": "tests/"
+          "Tests\\": "tests/"
         }
     }
 }
