+++ tests/TestConfigTemplate.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Zamzar;
+namespace Tests;
 
 // Rename to TestConfig.php on local machine and specify api key
 trait TestConfig
